refactor: move FieldOfView, ConicalFOV, CustomFOV, RectangularFOV into new fieldofview.py
