+++ gmatpyplus/hardware/__init__.py
@@ -1,3 +1,4 @@
+from .fieldofview import *
 from .fueltank import *
 from .hardware import *
 from .thruster import *
